requirements.txt
# V5.4 Tree-Based Training Dependencies
polars>=1.0
pyarrow>=15.0
scikit-learn>=1.5
shap>=0.45
pyyaml>=6.0
